--- src/client/wallet/WalletSummaryInfo/components/WalletActions.less
@@ -1,48 +1,0 @@
-@import (reference) '../../../styles/custom.less';
-
-.WalletAction {
-  display: flex;
-
-  &__button {
-    background: transparent;
-    border: 1px solid @orange-hot;
-    border-radius: 4px 0 0 4px;
-    color: @orange-hot;
-
-    &--withoutSelect {
-      border-radius: 4px;
-    }
-  }
-
-  &__select {
-    .ant-select-selection {
-      height: 30px;
-      background: transparent;
-      border: 1px solid @orange-hot;
-      border-radius: 0 4px 4px 0;
-      border-left: none;
-      outline: none;
-
-      &:focus,
-      &:hover {
-        outline: none;
-        box-shadow: none;
-      }
-
-      .anticon {
-        color: @orange-hot;
-      }
-    }
-  }
-
-  &--withoutButton {
-    .ant-select-selection {
-      border: 1px solid @orange-hot;
-      border-radius: 4px;
-    }
-  }
-
-  &__select-dropdown {
-    width: auto !important;
-  }
-}
